requirements.txt
asgiref==3.8.1
boto3==1.34.144
botocore==1.34.144
Django
django-cors-headers==4.4.0
django-filter==24.2
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
jmespath==1.0.1
pillow==10.4.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
urllib3==2.2.2
whitenoise==6.7.0
gunicorn==21.2.0
psycopg2-binary
django-environ
dj-database-url
django-storages
azure-storage-blob
